Cleanups from stricter span usage of size_t: media
More use of first(), compile-time sizes, etc.

Bug: 332941733
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6021592
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384563}

+++ media/base/decoder_buffer_unittest.cc
@@ -114,7 +114,7 @@ TEST(DecoderBufferTest, FromPlatformSharedMemoryRegion) {
 TEST(DecoderBufferTest, FromPlatformSharedMemoryRegion_Unaligned) {
   const uint8_t kData[] = "XXXhello";
   const size_t kDataSize = std::size(kData);
-  const off_t kDataOffset = 3;
+  const size_t kDataOffset = 3;
 
   auto region = base::UnsafeSharedMemoryRegion::Create(kDataSize);
   auto mapping = region.Map();
@@ -164,7 +164,7 @@ TEST(DecoderBufferTest, FromSharedMemoryRegion) {
 TEST(DecoderBufferTest, FromSharedMemoryRegion_Unaligned) {
   const uint8_t kData[] = "XXXhello";
   const size_t kDataSize = std::size(kData);
-  const off_t kDataOffset = 3;
+  const size_t kDataOffset = 3;
 
   auto mapping_region = base::ReadOnlySharedMemoryRegion::Create(kDataSize);
   ASSERT_TRUE(mapping_region.IsValid());
